final-main/config.py
# -*- coding: utf-8 -*-
import os
from dotenv import load_dotenv
load_dotenv()

basedir = os.path.abspath(os.path.dirname(__file__))

class Config:
    # Seguridad
    SECRET_KEY = os.environ["SECRET_KEY"]
    DEBUG = os.environ.get("FLASK_DEBUG", "TRUE") == "TRUE"
    TESTING = False



   